requirements.txt
python-docx==0.8.11
openai==0.28.1
PyQt5==5.15.9
tqdm==4.65.0
python-dotenv==1.0.0